README.md
Circle Search Auto-Clicker. The project is written in C#. A screen resolution of 1920x1080 (100% scaling) is required. The bot works by capturing screenshots of the screen and searching for circles with a customizable diameter. You can start the bot by pressing the **F** key, after which the bot will begin its work, and your cursor will always move to the detected circles on the screen, right-clicking on them. To stop the bot, press the **C** key.